Fix review findings

+++ lib/portrait_layout.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_master_detail_example/constants.dart';
 
+import 'constants.dart';
 import 'page_item.dart';
 
 class PortraitLayout extends StatefulWidget {
@@ -29,7 +29,7 @@ class _PortraitLayoutState extends State<PortraitLayout> {
         final page = widget.pages[index];
         return Scaffold(
           appBar: AppBar(
-            toolbarHeight: kAppBarHeight.toDouble(),
+            toolbarHeight: kAppBarHeight,
             title: Text('Portrait: ${page.title}'),
             leading: BackButton(
               onPressed: () {
@@ -56,7 +56,7 @@ class _PortraitLayoutState extends State<PortraitLayout> {
               builder: (context) {
                 return Scaffold(
                   appBar: AppBar(
-                    toolbarHeight: kAppBarHeight.toDouble(),
+                    toolbarHeight: kAppBarHeight,
                     title: Text('Portrait'),
                   ),
                   body: ListView(
